Added redo hotkey combination (Control-Y).
