+++ src/main/java/edu/wpi/cs3733/c20/teamS/Editing/tools/Memento.java
@@ -9,6 +9,7 @@ public abstract class Memento {
         execute();
     }
 
+
     public static Memento create(Runnable execute, Runnable undo, Runnable redo) {
         if (execute == null) ThrowHelper.illegalNull("execute");
         if (undo == null) ThrowHelper.illegalNull("undo");
